cryptfs: fix wrong password set by user during bootup
As per current implementation, When encryption is triggered from
settings, the scrypted intermediate key is created by using the
actual password but actual master key is encrypted with
default_password. Added changes to compare the current scrypted
intrmediate key with the stored intermediate key, so that the
password will be validated with actual password during bootup.

CRs-Fixed: 1075994
Change-Id: I2dd7112d46b89f53870b9f9fcd25061c321c9d22
Signed-off-by: Mohan CM <mohanmanjappa@gmail.com>